Move JWT decryption in auth file, add user role validation function
